catalog/requirements.txt
asgiref==3.8.1
attrs==25.3.0
colorama==0.4.6
coverage==7.8.2
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
django-ranged-response==0.2.0
django-simple-captcha==0.6.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
Faker==37.1.0
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
packaging==25.0
pillow==11.1.0
pluggy==1.6.0
PyJWT==2.9.0
pytest==8.3.5
pytest-cov==6.1.1
pytest-django==4.11.1
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
captcha==0.4.0